# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Velora DNS Management API",
    "version": "1.0.0",
    "description": "Session cookies require X-CSRF-Token on mutations. API bearer tokens bypass CSRF and require the documented scope. All API responses use a data or error envelope except text/dns zone export. Secrets are never returned by read endpoints."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080"
    }
  ],
  "paths": {
...
    "/api/v1/audit": {
      "get": {
        "operationId": "get_audit",
        "summary": "Get audit",
        "tags": [
          "audit"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AuditEvent"
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error; codes include authentication_required, insufficient_role, invalid_csrf, invalid_json, and endpoint-specific validation/storage codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "x-required-role": "admin"
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "post_auth_login",
        "summary": "Post auth/login",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Session"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error; codes include authentication_required, insufficient_role, invalid_csrf, invalid_json, and endpoint-specific validation/storage codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [],
        "x-required-role": "anonymous",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "writeOnly": true
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        }
      }
    },
...
    "/api/v1/backup/create": {
      "post": {
        "operationId": "post_backup_create",
        "summary": "Post backup/create",
        "tags": [
          "backup"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "default": {
            "description": "Error; codes include authentication_required, insufficient_role, invalid_csrf, invalid_json, and endpoint-specific validation/storage codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "x-required-role": "admin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "passphrase": {
                    "type": "string",
                    "format": "password",
                    "writeOnly": true,
                    "minLength": 12
                  }
                },
                "required": [
                  "passphrase"
                ]
              }
            }
          }
        }
      }
    },
    "/api/v1/backup/inspect": {
      "post": {
        "operationId": "post_backup_inspect",
        "summary": "Post backup/inspect",
        "tags": [
          "backup"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/BackupInspection"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error; codes include authentication_required, insufficient_role, invalid_csrf, invalid_json, and endpoint-specific validation/storage codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "x-required-role": "operator-or-admin",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "passphrase": {
                    "type": "string",
                    "description": "Must be the first part"
                  },
                  "bundle": {
                    "type": "string",
                    "format": "binary",
                    "description": "Encrypted .vdns bundle, at most 2 GiB"
                  }
                },
                "required": [
                  "passphrase",
                  "bundle"
                ]
              },
              "encoding": {
                "bundle": {
                  "contentType": "application/octet-stream"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backup/restore": {
      "get": {
        "operationId": "get_backup_restore",
        "summary": "Get backup/restore",
        "tags": [
          "backup"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/RestoreState"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error; codes include authentication_required, insufficient_role, invalid_csrf, invalid_json, and endpoint-specific validation/storage codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "x-required-role": "viewer-or-higher"
      },
      "post": {
        "operationId": "post_backup_restore",
        "summary": "Post backup/restore",
        "tags": [
          "backup"
        ],
        "responses": {
          "202": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/RestoreScheduled"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error; codes include authentication_required, insufficient_role, invalid_csrf, invalid_json, and endpoint-specific validation/storage codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "x-required-role": "operator-or-admin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  },
                  "confirm": {
                    "type": "boolean",
                    "description": "Must be true"
                  }
                },
                "required": [
                  "token",
                  "confirm"
                ]
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cache/entries": {
      "get": {
        "operationId": "get_cache_entries",
        "summary": "Get cache/entries",
        "tags": [
          "cache"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CacheEntryPage"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error; codes include authentication_required, insufficient_role, invalid_csrf, invalid_json, and endpoint-specific validation/storage codes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "x-required-role": "viewer-or-higher"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "velora_ API token; scopes read, write, or admin"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "velora_session"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Required for cookie-authenticated mutations"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "AuditEvent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time"
          },
          "actor": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "target": {
            "type": "string"
          },
          "result": {
            "type": "string"
          },
          "status_code": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "CacheEntryPage": {
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CacheEntry"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  }
}